refactor: split out `Section` and `Article` tags
